requirements.txt
Django==2.2.1
django-bootstrap-modal-forms==1.4.2
django-crispy-forms==1.7.2
django-rest-framework==0.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.4
gunicorn==19.9.0
pytz==2019.1
sqlparse==0.3.0
